requirements.txt
Flask==1.0.2
python-dotenv==0.9.1
flask-sqlalchemy==2.3.1
requests==2.19.1
Flask-Caching==1.4.0
flask-cors==3.0.6
image==1.5..27
Pillow==5.3.0
boto3==1.9.37 
pandas==0.23.4
mysqlclient==1.3.13
